np int update

+++ bb_tracking/track_generator.py
@@ -166,8 +166,8 @@ class TrackGenerator():
 
         if len(tracklet0_indices) > 0:
             all_distances = self.tracklet_cost_fn(all_features[:len(tracklet0_indices), :])
-            tracklet0_indices = np.array(tracklet0_indices, dtype=np.int32)
-            tracklet1_indices = np.array(tracklet1_indices, dtype=np.int32)
+            tracklet0_indices = np.array(tracklet0_indices, dtype=int)
+            tracklet1_indices = np.array(tracklet1_indices, dtype=int)
 
             tracklet_generator.fill_distance_matrix(tracklet0_indices, tracklet1_indices, all_distances, cost_matrix)
             track_indices, tracklet_indices = linear_sum_assignment(cost_matrix)
